# contracts/openapi/metadata-service.json
{
  "components": {
    "schemas": {
      "ActiveDataSource": {
        "properties": {
          "engine": {
            "title": "Engine",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "engine",
          "status",
          "last_sync_at"
        ],
        "title": "ActiveDataSource",
        "type": "object"
      },
      "ActiveDataSourceList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActiveDataSource"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ActiveDataSourceList",
        "type": "object"
      },
      "AgentContextColumn": {
        "properties": {
          "column_name": {
            "title": "Column Name",
            "type": "string"
          },
          "data_type": {
            "title": "Data Type",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "required": [
          "column_name",
          "data_type",
          "description"
        ],
        "title": "AgentContextColumn",
        "type": "object"
      },
      "AgentContextResponse": {
        "description": "What an agent may see of a data source: agent-visible tables, non-PII columns, no pointer\nto any credential.",
        "properties": {
          "allowed_schemas": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Schemas",
            "type": "array"
          },
          "data_source_id": {
            "format": "uuid",
            "title": "Data Source Id",
            "type": "string"
          },
          "engine": {
            "title": "Engine",
            "type": "string"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tables": {
            "items": {
              "$ref": "#/components/schemas/AgentContextTable"
            },
            "title": "Tables",
            "type": "array"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "data_source_id",
          "tenant_id",
          "engine",
          "status",
          "allowed_schemas",
          "last_sync_at",
          "tables"
        ],
        "title": "AgentContextResponse",
        "type": "object"
      },
      "AgentContextTable": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/AgentContextColumn"
            },
            "title": "Columns",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "row_count_estimate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count Estimate"
          },
          "schema_name": {
            "title": "Schema Name",
            "type": "string"
          },
          "table_name": {
            "title": "Table Name",
            "type": "string"
          }
        },
        "required": [
          "schema_name",
          "table_name",
          "description",
          "row_count_estimate",
          "columns"
        ],
        "title": "AgentContextTable",
        "type": "object"
      },
...
      "DataSourceCreateRequest": {
        "additionalProperties": false,
        "description": "Non-secret connection metadata. Credentials go to `POST .../secret`.",
        "properties": {
          "allowed_schemas": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Allowed Schemas",
            "type": "array"
          },
          "database_name": {
            "maxLength": 63,
            "minLength": 1,
            "title": "Database Name",
            "type": "string"
          },
          "engine": {
            "enum": [
              "postgres",
              "mysql",
              "snowflake",
              "bigquery",
              "redshift"
            ],
            "title": "Engine",
            "type": "string"
          },
          "host_label": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Host Label",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "engine",
          "host_label",
          "database_name",
          "allowed_schemas"
        ],
        "title": "DataSourceCreateRequest",
        "type": "object"
      },
      "DataSourceListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DataSourceResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "DataSourceListResponse",
        "type": "object"
      },
      "DataSourceResponse": {
        "properties": {
          "allowed_schemas": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Schemas",
            "type": "array"
          },
          "capabilities": {
            "additionalProperties": true,
            "title": "Capabilities",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "format": "uuid",
            "title": "Created By",
            "type": "string"
          },
          "database_name": {
            "title": "Database Name",
            "type": "string"
          },
          "engine": {
            "title": "Engine",
            "type": "string"
          },
          "host_label": {
            "title": "Host Label",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "name",
          "engine",
          "host_label",
          "database_name",
          "allowed_schemas",
          "capabilities",
          "status",
          "last_sync_at",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "DataSourceResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QueryPolicyColumn": {
        "properties": {
          "column_name": {
            "title": "Column Name",
            "type": "string"
          },
          "data_type": {
            "title": "Data Type",
            "type": "string"
          },
          "is_pii": {
            "title": "Is Pii",
            "type": "boolean"
          }
        },
        "required": [
          "column_name",
          "data_type",
          "is_pii"
        ],
        "title": "QueryPolicyColumn",
        "type": "object"
      },
      "QueryPolicyResponse": {
        "description": "Everything query-gateway needs to validate and route a query. A pointer, never a secret.",
        "properties": {
          "allowed_schemas": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Schemas",
            "type": "array"
          },
          "data_source_id": {
            "format": "uuid",
            "title": "Data Source Id",
            "type": "string"
          },
          "database_name": {
            "title": "Database Name",
            "type": "string"
          },
          "engine": {
            "title": "Engine",
            "type": "string"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "secret_ref": {
            "title": "Secret Ref",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tables": {
            "items": {
              "$ref": "#/components/schemas/QueryPolicyTable"
            },
            "title": "Tables",
            "type": "array"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "data_source_id",
          "tenant_id",
          "engine",
          "database_name",
          "allowed_schemas",
          "status",
          "secret_ref",
          "last_sync_at",
          "tables"
        ],
        "title": "QueryPolicyResponse",
        "type": "object"
      },
      "QueryPolicyTable": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/QueryPolicyColumn"
            },
            "title": "Columns",
            "type": "array"
          },
          "is_visible_to_agent": {
            "title": "Is Visible To Agent",
            "type": "boolean"
          },
          "schema_name": {
            "title": "Schema Name",
            "type": "string"
          },
          "table_name": {
            "title": "Table Name",
            "type": "string"
          }
        },
        "required": [
          "schema_name",
          "table_name",
          "is_visible_to_agent",
          "columns"
        ],
        "title": "QueryPolicyTable",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Data-source connection metadata, the schema catalog and catalog sync for the BuVi platform (build spec Sections 8.2, 13.1).",
    "title": "metadata-service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/data-sources": {
      "get": {
        "description": "Tenant-scoped list (Section 9: `data:manage`).",
        "operationId": "list_data_sources_api_v1_data_sources_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSourceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Data Sources",
        "tags": [
          "data-sources"
        ]
      },
      "post": {
        "description": "Create a connection record; `pending` until credentials are set (Section 9).",
        "operationId": "create_data_source_api_v1_data_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataSourceCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSourceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Data Source",
        "tags": [
          "data-sources"
        ]
      }
    },
...
    "/internal/v1/data-sources": {
      "get": {
        "operationId": "active_data_sources_internal_v1_data_sources_get",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveDataSourceList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Active Data Sources",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/v1/data-sources/{data_source_id}/context": {
      "get": {
        "operationId": "agent_context_internal_v1_data_sources__data_source_id__context_get",
        "parameters": [
          {
            "in": "path",
            "name": "data_source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Data Source Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Agent Context",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/v1/data-sources/{data_source_id}/query-policy": {
      "get": {
        "operationId": "query_policy_internal_v1_data_sources__data_source_id__query_policy_get",
        "parameters": [
          {
            "in": "path",
            "name": "data_source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Data Source Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryPolicyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query Policy",
        "tags": [
          "internal"
        ]
      }
    }
  }
}